pyproject.toml
[tool.poetry]
name = "repl_python3_Phantom-Bot"
version = "0.1.0"
description = "A Clash of Clans Bot for Discord"
authors = ["Tyler Bryk"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
"coc.py" = "^1.3.0"
pandas = "^1.3.0"
"discord.py" = "^1.7.3"
PyMySQL = "^1.0.2"
asyncpraw = "^7.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
